refactor(models): rename model files from plural to singular and update imports
